backend/requirements.txt
asgiref==3.11.1
channels==4.3.2
channels_redis==4.3.0
Django==6.0.4
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
pypdf==6.8.0
groq==1.1.0
python-docx==1.2.0
python-pptx==1.0.2
Pillow==12.1.1
pytesseract==0.3.13
python-dotenv==1.2.2
redis==7.4.0
sqlparse==0.5.5
tzdata==2026.1
numpy==2.4.4
pandas==3.0.2
python-dateutil==2.9.0.post0
six==1.17.0
